Dont play animation when undo
